refactor(tui): extract the agents overlay into tui::render::agents
Panel geometry, card formatting, the working-glyph animation, and the
overlay help line move out of the module root with their tests. The
agent-roster fixtures join tui::test_support.
