+++ src/tui/render/mod.rs
@@ -1,4 +1,5 @@
 //! Everything that draws: chrome, overlays, modals, and pane contents.
 
+pub(in crate::tui) mod agents;
 pub(in crate::tui) mod footer;
 pub(in crate::tui) mod vt;
